Crate wtf8

Implementation of the WTF-8 encoding.

This library uses Rust’s type system to maintain well-formedness, like the String and &str types do for UTF-8.

Since WTF-8 must not be used for interchange, this library deliberately does not provide access to the underlying bytes of WTF-8 strings, nor can it decode WTF-8 from arbitrary bytes. WTF-8 strings can be obtained from UTF-8, UTF-16, or code points.

Structs

CodePoint
A Unicode code point: from U+0000 to U+10FFFF.

IllFormedUtf16CodeUnits

Wtf8
A borrowed slice of well-formed WTF-8 data.

Wtf8Buf
An owned, growable string of well-formed WTF-8 data.

Wtf8CodePoints
Iterator for the code points of a WTF-8 string.